+++ src/lib/memberIdBundle/MemberIdBundleSchema.ts
@@ -2,7 +2,7 @@
 import { AsnProp } from '@peculiar/asn1-schema';
 import { Certificate } from '@peculiar/asn1-x509';
 
-import { DnssecChain } from './dns/DnssecChain.js';
+import { DnssecChain } from '../dns/DnssecChain.js';
 
 export class MemberIdBundleSchema {
   @AsnProp({ type: DnssecChain })
